Split tests on smoke and extended + separate tests for valid and invalid input parameters
