+++ tests_ollm/extended_tests.py
@@ -1,0 +1,229 @@
+import os
+import types
+import time
+from pathlib import Path
+
+import requests
+import pytest
+
+
+import numpy as np
+from scipy.spatial import distance
+from concurrent.futures import ThreadPoolExecutor
+import concurrent.futures
+
+import openai
+
+from utils import run_chat_completion, send_request_get_response, send_request_with_timeout, path_to_file
+
+# For compatibility with OpenAI versions before v1.0
+# https://github.com/openai/openai-python/pull/677.
+OPENAI_VER_MAJ = int(openai.__version__.split(".")[0])
+
+if OPENAI_VER_MAJ >= 1:
+    from openai import APIError, AuthenticationError, APIConnectionError
+    from pydantic import BaseModel as CompletionObject
+else:
+    from openai.error import APIError, AuthenticationError, APIConnectionError
+    from openai.openai_object import OpenAIObject as CompletionObject
+
+from sentence_transformers import SentenceTransformer
+
+@pytest.fixture
+def model_name(request):
+    return request.config.getoption("--model_name")
+
+
+@pytest.fixture
+def token():
+    return os.environ["OCTOAI_TOKEN"]
+
+
+@pytest.fixture
+def endpoint(request):
+    return request.config.getoption("--endpoint")
+
+@pytest.fixture
+def context_size(request):
+    return request.config.getoption("--context_size", default=4096)
+
+def test_cancel_and_follow_up_requests(model_name, token, endpoint):
+    data = {
+        "model": model_name,
+        "messages": [
+            {
+                "role": "user",
+                "content": "Create a big story about a friendship between a cat and a dog.",
+            }
+        ],
+        "max_tokens": 500,
+        "n": 1,
+        "stream": False,
+        "stop": None,
+        "temperature": 0.8,
+        "top_p": 1.0,
+        "presence_penalty": 0,
+        "return_completion": False,
+    }
+    url = endpoint + "/v1/chat/completions"
+    headers = {
+        "Content-Type": "application/json",
+        "Authorization": f"Bearer {token}",
+    }
+    try:
+        requests.post(url, json=data, headers=headers, timeout=1)
+    except requests.exceptions.Timeout:
+        print("Timeout of request")
+
+    follow_up_request = requests.post(url, json=data, headers=headers).json()
+    assert "created" in follow_up_request
+
+
+def test_canceling_requests(model_name, token, endpoint):
+    data = {
+        "model": model_name,
+        "messages": [
+            {
+                "role": "user",
+                "content": "Create a big story about a friendship between a cat and a dog.",
+            }
+        ],
+        "max_tokens": 1000,
+        "n": 1,
+        "stream": False,
+        "stop": None,
+        "temperature": 0.0,
+        "top_p": 1.0,
+        "presence_penalty": 0,
+        "return_completion": False,
+    }
+    url = endpoint + "/v1/chat/completions"
+    headers = {
+        "Content-Type": "application/json",
+        "Authorization": f"Bearer {token}",
+    }
+    num_workers = 8
+    responses_code_set = set()
+
+    start_time = time.time()
+    with ThreadPoolExecutor(max_workers=num_workers) as executor:
+        futures = [executor.submit(send_request_get_response, url, data, headers) for _ in range(num_workers)]
+        for future in concurrent.futures.as_completed(futures):
+            responses_code_set.add(future.result().status_code)
+    first_run_time = time.time() - start_time
+    assert (responses_code_set == {200}), f"There is a problem with sending request"
+    
+    with ThreadPoolExecutor(max_workers=num_workers) as executor:
+        futures = [executor.submit(send_request_with_timeout, url, data, headers) for _ in range(num_workers)]
+
+    start_time = time.time()
+    with ThreadPoolExecutor(max_workers=num_workers) as executor:
+        futures = [executor.submit(send_request_get_response, url, data, headers) for _ in range(num_workers)]
+        for future in concurrent.futures.as_completed(futures):
+            responses_code_set.add(future.result().status_code)
+    second_run_time = time.time() - start_time
+    assert (responses_code_set == {200}), f"There is a problem with sending request"
+
+    print(first_run_time, second_run_time)
+    threshold = 5
+    assert abs(second_run_time - first_run_time) < threshold
+
+
+@pytest.mark.parametrize("temperature", [0.0, 0.5, 0.7, 1.0, 1.5])
+def test_same_completion_len(temperature, model_name, token, endpoint):
+    messages = [
+        {"role": "user", "content": "Hello, how can you help me? Answer short."}
+    ]
+    tokens_arr = []
+    mean = 0
+    trials = 4
+    for _ in range(trials):
+        completion = run_chat_completion(
+            model_name,
+            messages,
+            token,
+            endpoint,
+            temperature=temperature,
+            top_p=1.0,
+            return_completion=True,
+        )
+        mean += completion["usage"]["completion_tokens"]
+        tokens_arr.append(completion["usage"]["completion_tokens"])
+
+    mean /= trials
+    threshold = 10
+    assert all([abs(tokens_arr[i] - mean) <= threshold for i in range(trials)])
+
+
+def test_multiple_messages(model_name, token, endpoint):
+    messages = [
+        {
+            "role": "user",
+            "content": "What is the capital of France?",
+        },
+        {
+            "role": "assistant",
+            "content": "Paris",
+        },
+        {
+            "role": "user",
+            "content": "2 + 2 =",
+        },
+    ]
+
+    completion = run_chat_completion(
+        model_name, messages, token, endpoint, max_tokens=20, return_completion=True
+    )
+    assert "4" in completion["choices"][0]["message"]["content"]
+
+
+@pytest.mark.parametrize("input_tokens", [496, 963, 2031, 3119, 3957, 5173])
+def test_large_input_content(input_tokens, model_name, context_size, token, endpoint):
+    with open(path_to_file(f"input_context/text_about_{input_tokens}_tokens.txt"), "r") as file:
+        prompt = file.read()
+    messages = [
+        {"role": "user", "content": prompt}
+    ]
+    max_tokens = 200
+    if model_name == "codellama-34b-instruct-fp16":
+        context_size = 16384
+        
+    if (input_tokens + max_tokens) < context_size: 
+        assert run_chat_completion(model_name, messages, token, endpoint, max_tokens=max_tokens) == 200
+    else:
+        assert run_chat_completion(model_name, messages, token, endpoint, max_tokens=max_tokens) == 400
+
+
+def test_send_many_request(model_name, token, endpoint):
+    data = {
+        "model": model_name,
+        "messages": [
+            {
+                "role": "user",
+                "content": "Create a short story about a friendship between a cat and a dog.",
+            }
+        ],
+        "max_tokens": 300,
+        "n": 1,
+        "stream": False,
+        "stop": None,
+        "temperature": 0.0,
+        "top_p": 1.0,
+        "presence_penalty": 0,
+        "return_completion": False,
+    }
+
+    url = endpoint + "/v1/chat/completions"
+    headers = {
+        "Content-Type": "application/json",
+        "Authorization": f"Bearer {token}",
+    }
+    responses_code_set = set()
+    num_workers = 64
+
+    with ThreadPoolExecutor(max_workers=num_workers) as executor:
+        futures = [executor.submit(send_request_get_response, url, data, headers) for _ in range(num_workers)]
+        for future in concurrent.futures.as_completed(futures):
+            responses_code_set.add(future.result().status_code)
+
+    assert responses_code_set == {200}
